Break larger part of network-audit out into separate file
